+++ docs/modules/airflow/examples/example-pyspark-pi.yaml
@@ -3,17 +3,10 @@ apiVersion: spark.stackable.tech/v1alpha1
 kind: SparkApplication
 metadata:
   name: pyspark-pi
-  namespace: default
 spec:
-  version: "1.0"
-  sparkImage: docker.stackable.tech/stackable/pyspark-k8s:3.3.0-stackable0.0.0-dev
+  sparkImage:
+    productVersion: 3.5.0
   mode: cluster
   mainApplicationFile: local:///stackable/spark/examples/src/main/python/pi.py
-  driver:
-    cores: 1
-    coreLimit: "1200m"
-    memory: "512m"
   executor:
-    cores: 1
-    instances: 3
-    memory: "512m"
+    replicas: 1
