+++ docs/modules/airflow/pages/usage-guide/applying-custom-resources.adoc
@@ -1,10 +1,10 @@
 = Applying Custom Resources
 
-Airflow can be used to apply custom resources from within a cluster. An example of this could be a SparkApplication job that is to be triggered by Airflow. The steps below describe how this can be done.
+Airflow can be used to apply custom resources from within a cluster. An example of this could be a SparkApplication job that is to be triggered by Airflow. The steps below describe how this can be done. The DAG will consist of modularized python files and will be provisioned using the git-sync facility.
 
 == Define an in-cluster Kubernetes connection
 
-An in-cluster connection can either be created from within the Webserver UI (note that the "in cluster configuration" box is ticked):
+To start a Spark job, Airflow needs to be able to communicate with Kubernetes and an in-cluster connection is required for this, which can be created from within the Webserver UI (note that the "in cluster configuration" box is ticked):
 
 image::airflow_connection_ui.png[Airflow Connections]
 
@@ -38,19 +38,41 @@ include::example$example-airflow-spark-clusterrolebinding.yaml[]
 
 == DAG code
 
-Now for the DAG itself. The job to be started is a simple Spark job that calculates the value of pi:
+Now for the DAG itself. The job to be started is a modularized DAG that uses starts a one-off Spark job that calculates the value of pi. The file structure fetched to the root git-sync folder looks like this:
+
+----
+dags
+|_ stackable
+  |_ __init__.py
+  |_ spark_kubernetes_operator.py
+  |_ spark_kubernetes_sensor.py
+|_ pyspark_pi.py
+|_ pyspark_pi.yaml
+----
+
+The Spark job will calculate the value of pi using one of the example scripts that comes bundled with Spark:
 
 [source,yaml]
 ----
 include::example$example-pyspark-pi.yaml[]
 ----
 
-This will called from within a DAG by using the connection that was defined earlier. It will be wrapped by the `KubernetesHook` that the Airflow Kubernetes provider makes available [here.](https://github.com/apache/airflow/blob/main/airflow/providers/cncf/kubernetes/operators/spark_kubernetes.py) There are two classes that are used to:
+This will be called from within a DAG by using the connection that was defined earlier. It will be wrapped by the `KubernetesHook` that the Airflow Kubernetes provider makes available https://github.com/apache/airflow/blob/main/airflow/providers/cncf/kubernetes/operators/spark_kubernetes.py[here].There are two classes that are used to:
 
 - start the job
 - monitor the status of the job
 
-These are written in-line in the python code below, though this is just to make it clear how the code is used (the classes `SparkKubernetesOperator` and `SparkKubernetesSensor` will be used for all custom resources and thus are best defined as separate python files that the DAG would reference).
+The classes `SparkKubernetesOperator` and `SparkKubernetesSensor` are in dedicated files as they will typically be used for all custom resources and thus are best defined separately from that the DAG that calls them. This also demonstrates that modularized DAG structures can be used for defining jobs as long as all dependencies exist in or below the root folder pulled by git-sync.
+
+[source,python]
+----
+include::example$example_spark_kubernetes_operator.py[]
+----
+
+[source,python]
+----
+include::example$example_spark_kubernetes_sensor.py[]
+----
 
 [source,python]
 ----
@@ -71,3 +93,7 @@ image::airflow_dag_graph.png[Airflow Connections]
 Clicking on the "spark_pi_monitor" task and selecting the logs shows that the status of the job has been tracked by Airflow:
 
 image::airflow_dag_log.png[Airflow Connections]
+
+NOTE: If the `KubernetesExecutor` is employed the logs are only accessible via the SDP logging mechanism, described https://docs.stackable.tech/home/stable/concepts/logging[here].
+
+TIP: A full example of the above is used as an integration test https://github.com/stackabletech/airflow-operator/tree/main/tests/templates/kuttl/mount-dags-gitsync[here].
