+++ tests/templates/kuttl/mount-dags-gitsync/dags/pyspark_pi.yaml
@@ -8,28 +8,5 @@ spec:
     productVersion: 3.5.0
   mode: cluster
   mainApplicationFile: local:///stackable/spark/examples/src/main/python/pi.py
-  job:
-    config:
-      resources:
-        cpu:
-          min: 500m
-          max: 500m
-        memory:
-          limit: 512Mi
-  driver:
-    config:
-      resources:
-        cpu:
-          min: 1000m
-          max: 1200m
-        memory:
-          limit: 1024Mi
   executor:
-    config:
-      resources:
-        cpu:
-          min: 500m
-          max: 1000m
-        memory:
-          limit: 1024Mi
     replicas: 1
